sys/newlib_syscalls_default: add a function to initialize the heap
In our port of RIOT on Pip for the Cortex-M CPUs, Pip and RIOT are
compiled separately. Only Pip knows the memory layout of the target.
RIOT is seen as a binary that can be loaded at an arbitrary address in
memory. Therefore, during the compilation of RIOT, the linker does not
know the start and end addresses of the heap. These values will be
passed as parameters by Pip at the beginning of the execution of RIOT.