cpu/nrf52_pip: update of the Pip_in() function call
Previously, the Pip_in() function stored the value of the read register
in the second argument of the function and returned a value to the
calling partition, indicating whether the read/write had failed.

Now, the value of the read register is returned to the calling partition
instead of an error code. In case of error, Pip trigger a MemManage
fault, which is propagated to the parent partition.