pkg/newlib_nano_pip: create the newlib_nano_pip package
RIOT uses by default the newlib which is within the Arm GNU Toolchain of
the host. This does not work with our port of RIOT on Pip because the
newlib has not been compiled with the position independent code flags.

This package clones the official newlib repository and compiles it (only
once per architecture) with the appropriate position independent code
flags.