cpu/cortexm_common_pip/ldscripts: move the .roxfa input section
The auto_init() function calls the _auto_init_module() function on each
module of the auto-initialization array. The _auto_init_module()
function branches to the address of the module's initialization
function, contained in the init field of the auto_init_module_t
structure.

The problem is that the auto-initialization array is placed in the
.roxfa input section, which is itself placed in the .rom output section.
Thus, the init field contains an erroneous address, since the addresses
in the .rom output section cannot be relocated by the crt0.

To correct the problem, simply move the .roxfa input section into the
.rom.ram output section.